Refactored wallet module logic to classes.
