+++ packages/core/src/lib/services/index.ts
@@ -6,7 +6,6 @@ export * from "./storage/json-storage.service";
 export * from "./storage/web-storage.service";
 export * from "./logger/logger.service";
 export * from "./logger/logger.service.types";
-// export * from "./logger/logging.service";
-// export * from "./logger/logging.service.types";
+export * from "./wallet-modules/wallet-modules.service";
 export * from "./event-emitter/event-emitter.service";
 export * from "./event-emitter/event-emitter.types";
