+++ packages/core/src/lib/wallet-selector.ts
@@ -6,9 +6,8 @@ import {
   WalletSelectorParams,
 } from "./wallet-selector.types";
 import { WalletSelectorModal } from "./modal/modal.types";
-import { EventEmitter, Logger } from "./services";
+import { EventEmitter, Logger, WalletModules } from "./services";
 import { Wallet } from "./wallet";
-import { setupWalletModules } from "./modules/wallet-modules";
 import { setupModal } from "./modal/modal";
 
 export const setupWalletSelector = async (
@@ -19,13 +18,15 @@ export const setupWalletSelector = async (
 
   const emitter = new EventEmitter<WalletSelectorEvents>();
   const store = await createStore(storage);
-  const walletModules = await setupWalletModules({
+  const walletModules = new WalletModules({
     factories: params.modules,
     storage,
     options,
     store,
     emitter,
   });
+
+  await walletModules.setup();
 
   // TODO: Remove omit once modal is a separate package.
   const selector: Omit<WalletSelector, keyof WalletSelectorModal> = {
